% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_contrasts}
\alias{check_contrasts}
\alias{\alias{check_contrasts}}
\title{function to check if a set of contrasts for a variable are orthogonal and consistent}
\usage{
check_contrasts(variable)
}
\arguments{
\item{variable}{}
}
\value{
the function returns an object of class 'pml' of the 'phangorn'
package. Advanced and elaborated plots can be drawn in later steps based on
the tree data of the pml class object
}
\description{
this function is a wrapper around the functions
\code{'\link[seqinr]{dist.alignment}'}, \code{'\link[ape]{dist.dna}'},
\code{'\link[ape]{nj}'}, \code{'\link[ape]{bionj}'},
\code{'\link[ape]{fastme.bal}'}, \code{'\link[ape]{fastme.ols}'},
\code{'\link[phangorn]{pml}'} and \code{'\link[phangorn]{optim.pml}'}. it takes a sequences alignment in
format 'alignment' of 'DNAbin' matrix and perform all transformations and steps
to calculate a phylogenetic distance matrix based on similarity or identity
in the case of proteins or based in evolutionary models in the case of DNA or
RNA, to perform a likelihood-based phylogenetic clustering and to optimise the phylogeny by a
maximum likelihood algorithm
}
\examples{
data(fastaRNA)
data(phylipProt)
mytree <- max_likelihood(fastaRNA, type=RNA, clustering=fastme.bal,
                        pml.model=GTR, clean=FALSE)
\dontrun{mytree <- max_likelihood(phylipProt, type=protein, pml.model=Blosum62,
                        outgroup=YP_0010399)}
\dontrun{plot.phylo(mytree, type='u')}

}
\seealso{
\code{'\link{retrieve_seqs}'} \code{'\link{print_alignment}'}
\code{'\link{clean_alignment}'} \code{'\link{load_alignment}'}
\code{'\link{make_tree}'} \code{'\link{max_parsimony}'}
}
\author{
gerardo esteban antonicelli
}
