% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corr.test}
\alias{corr.test}
\alias{\alias{corr.test}}
\title{function to calculate correlation coefficients, confidence intervals and
p-values for one or more pairs of variables}
\usage{
corr.test(
  data,
  method = "pearson",
  alternative = "two.sided",
  conf.level = 0.95,
  use = "pairwise.complete.obs",
  boots = 2000,
  verbose = TRUE,
  exact = NULL,
  continuity = FALSE
)
}
\arguments{
\item{data}{a character string without '' specifying a data.frame object with
the data, each variable must be in only one column, or a subset of a
data.frame with two or more variables or two vectors in a list}

\item{method}{a character string specifying the correlation method to be
used, options are pearson, spearman or kendall, names can be abbreviated}

\item{alternative}{a character string indicating the alternative hypothesis
and must be one of 'two.sided', 'greater' or 'less', 'greater' corresponds
to a positive association, 'less' to a negative association, names can be
abbreviated}

\item{conf.level}{numeric an decimal specifying the confidence level for the
returned confidence interval}

\item{use}{an character string giving a method for computing covariances in
the presence of missing values, options are (an abbreviation of) one of the
strings 'complete.obs' or 'pairwise.complete.obs' (default)}

\item{boots}{a single positive integer indicating the number of bootstrap
replicates, default=2000}

\item{verbose}{a logical indicating whether a summary of the output should be
printed to the console}

\item{exact}{a logical indicating whether an exact p-value should be computed
(used for Kendall's tau and Spearman's rho), see 'details' for the meaning
of NULL (the default)}

\item{continuity}{logical: if true, a continuity correction is used for
Kendall's tau and Spearman's rho when not computed exactly}
}
\value{
if the input are two vectors or two columns of a data.frame and the
method is pearson the function returns invisibly and print to the console
the output of cor.test(), if the method is kendall or spearman the function
returns invisibly a list and prints to the console (if verbose=TRUE) the
output of cor.test() and the confidence interval from boot.ci(), if the
input is a data.frame with more than two variables the function returns
invisibly a list with the call, relevant input parameters and the matrices
with correlation coefficients, p-values and lower and upper limits of
confidence intervals for all combinations of variables, if verbose=TRUE the
function prints to the console the matrices with correlation coefficients,
p-values and lower and upper limits of confidence intervals for all
combinations of variables
}
\description{
this function is an extension of cor.test() that calculates correlation
coefficients, p-values and confidence intervals for a pair of variables or
for all pairs of variables if more than two are provided in a data.frame. It
can calculate Pearson, Kendall or Spearman correlation coefficients and
p-values for a two-tailed or one-tailed alternative hypothesis. For Kendall
or Spearman correlation, confidence intervals are calculated by bootstrap
samling. In case of NAs the complete row can be taken out or only for the
affected comparisons
}
\details{
the three methods each estimate the association between paired
samples and compute a test of the value being zero. They use different
measures of association, all in the range \link{-1, 1} with 0 indicating no
association, these are sometimes referred to as tests of no correlation,
but that term is often confined to the default method. If method is
'pearson', the test statistic is based on Pearson's product moment
correlation coefficient cor(x, y) and follows a t distribution with
length(x)-2 degrees of freedom if the samples follow independent normal
distributions. If there are at least 4 complete pairs of observation, an
asymptotic confidence interval is given based on Fisher's Z transform. If
method is 'kendall' or 'spearman', Kendall's tau or Spearman's rho
statistic is used to estimate a rank-based measure of association, these
tests may be used if the data do not necessarily come from a bivariate
normal distribution. For Kendall's test, by default (if exact is NULL), an
exact p-value is computed if there are less than 50 paired samples
containing finite values and there are no ties, otherwise, the test
statistic is the estimate scaled to zero mean and unit variance, and is
approximately normally distributed. For Spearman's test, p-values are
computed using algorithm AS 89 for n<1290 and exact=TRUE, otherwise via the
asymptotic t approximation. Note that these are ‘exact’ for n<10, and use
an Edgeworth series approximation for larger sample sizes (the cutoff has
been changed from the original paper)
}
\examples{
data(examData)
data(personalityData)
data(stalkerData)
corr.test(examData[ , c(2, 4)], 'pearson', 'greater', 0.99)
corr.test(list(examData$Revise, examData$Anxiety), 'kendall', 'two.sided', 0.99, boots=1000)
corr.test(examData[ , 2:4], 'spearman', 'less', 0.95, 'complete.obs', exact=FALSE, verbose=FALSE)
corr.test(personalityData[ , c(3, 12)], conf.level=0.99)
corr.test(stalkerData[ , c(2, 3)], 'spearman', conf.level=0.99, exact=FALSE, continuity=TRUE)

}
\seealso{
\code{'\link{check_contrasts}'} \code{'\link{omega_factorial}'}
}
\author{
gerardo esteban antonicelli
}
