% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limma_fit}
\alias{limma_fit}
\alias{\alias{fit_aov}}
\title{function to retrieve the contrasts stored in a dataset and fit an ANOVA on
those}
\usage{
limma_fit(
  object,
  design = NULL,
  ndups = 1,
  spacing = 1,
  block = NULL,
  correlation,
  weights = NULL,
  method = "ls",
  ...
)
}
\arguments{
\item{anova}{an object of class anova data.frame with the ANOVA model created
by aov(), in the original dataset each variable must be in only one
column, and all contrasts for each variable should be loaded as contrasts
for that variable}
}
\value{
the function returns the fitted ANOVA models for all the contrasts stored for each variable
}
\description{
this function is a wrapper and an adapter around \code{'\link[stats]{aov}'},
it retrieves the contrasts for each independent variable specified in an
aov() model and stored in the original dataset, and fits an ANOVA model on
all contrasts for each independent variable, in order to obtained the
computed SS and MS for further calculations, otherwise not available from the
original ANOVA model as displayed by summary() or summary.lm(), the model
fitted for all contrasts of one variable is the same as displayed by
summary.lm() on the original ANOVA model
}
\examples{
data(gogglesData)
data(depressionData)
goggles.model <- aov(attractiveness ~ gender + alcohol + gender:alcohol, data=gogglesData)
simple.model <- aov(attractiveness ~ simple, data=gogglesData)
depression.model <- aov(diff ~ treat, data=depressionData)
fit_aov(goggles.model)
fit_aov(simple.model)
fit_aov(depression.model)

}
\seealso{
\code{'\link{check_contrasts}'} \code{'\link{omega_factorial}'}
}
\author{
gerardo esteban antonicelli
}
