% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{omega}
\alias{omega}
\alias{\alias{omega}}
\title{function to calculate the omega squared and f squared effect size of independent
variables in an ANOVA model}
\usage{
omega(anova)
}
\arguments{
\item{anova}{an object of class anova data.frame with the ANOVA model created
by aov(), in the original dataset each variable must be in only one
column, and all contrasts for each variable should be loaded as contrasts
for that variable}
}
\value{
the function returns the omega squared and f squared effect sizes for
each independent variable specified in the main ANOVA model and for each
contrast in the ANOVA model fitted for each independent variable
}
\description{
this function uses the function fit_aov() to retrieve the contrasts for each
independent variable specified in an aov() model and stored in the original
dataset, and to fit an ANOVA model on all contrasts for each independent
variable, in order to obtained the computed SS and MS for further calculations,
otherwise not available from the original ANOVA model as displayed by summary()
or summary.lm(), the model fitted for all contrasts of one variable is the same
as displayed by summary.lm() on the original ANOVA model
this function calculates the omega squared and f squared coefficients from the SS,
MS and df calculated by aov() for each independent variable and for each contrast
specified for each independent variable
}
\examples{
data(gogglesData)
data(depressionData)
goggles.model <- aov(attractiveness ~ gender + alcohol + gender:alcohol, data=gogglesData)
simple.model <- aov(attractiveness ~ simple, data=gogglesData)
depression.model <- aov(diff ~ treat, data=depressionData)
omega(goggles.model)
omega(simple.model)
omega(depression.model)

}
\seealso{
\code{'\link{check_contrasts}'} \code{'\link{omega_factorial}'}
}
\author{
gerardo esteban antonicelli
}
