% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{omega_factorial}
\alias{omega_factorial}
\alias{\alias{omega_factorial}}
\title{function to calculate the effect size of a comparison}
\usage{
omega_factorial(anova)
}
\arguments{
\item{anova}{an object of class anova data.frame with the results of an ANOVA
type III returned by Anova(model, type='III'), the model is created by
aov()}
}
\value{
the function prints the omega squared for two factors and one
interaction
}
\description{
function to calculate the effect size of a comparison
}
\examples{
data(anova)
omega_factorial(anova)

}
\seealso{
\code{'\link{check_contrasts}'} \code{'\link{es}'}
}
\author{
gerardo esteban antonicelli
}
