% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{unwrap_dups}
\alias{unwrap_dups}
\alias{\alias{unwrap_dups}}
\title{function to report which coefficients in a linear model cannot be estimated}
\usage{
unwrap_dups(M, ndups = 2, spacing = 1)
}
\arguments{
\item{x}{a numeric matrix or vector}
}
\value{
the function returns a character vector of names for the columns of x
which are linearly dependent on previous columns. If x has full column
rank, then the value is NULL
}
\description{
the function nonEstimable() from the limma package is used by lmFit() to
report which coefficients in a linear model cannot be estimated, the function
nonEst() in the GLM package is identical to nonEstimable() and it's used by
limma_fit()
}
\examples{
data(gogglesData)

}
\seealso{
\code{'\link{check_contrasts}'} \code{'\link{omega_factorial}'}
}
\author{
gerardo esteban antonicelli
}
